command migration to better system
