+++ docs/README.md
@@ -1,4 +1,4 @@
-# MikuBot User Guide
+# miku.MikuBot User Guide
 
 // Update the title above to match the actual product name
