+++ src/main/java/miku/command/CustomConsumer.java
@@ -1,3 +1,5 @@
+package miku.command;
+
 @FunctionalInterface
 public interface CustomConsumer<T> {
     void accept(T t) throws Exception;
