+++ src/main/java/miku/exceptions/IllegalSaveException.java
@@ -1,3 +1,5 @@
+package miku.exceptions;
+
 public class IllegalSaveException extends Exception {
     public IllegalSaveException(String str) {
         super(str);
