+++ src/main/java/miku/exceptions/UnrecognizedCommandException.java
@@ -1,3 +1,5 @@
+package miku.exceptions;
+
 public class UnrecognizedCommandException extends Exception {
     public UnrecognizedCommandException() {
         super("Miku doesn't recognize this command :(");
