+++ src/main/java/miku/storage/SaveDataManager.java
@@ -1,3 +1,8 @@
+package miku.storage;
+
+import miku.exceptions.IllegalSaveException;
+import miku.tasks.Task;
+
 import java.io.File;
 import java.io.IOException;
 import java.io.FileWriter;
