+++ src/main/java/miku/tasks/DeadlineTask.java
@@ -1,3 +1,5 @@
+package miku.tasks;
+
 public class DeadlineTask extends Task {
     private String deadline;
