+++ src/main/java/miku/tasks/Task.java
@@ -1,3 +1,5 @@
+package miku.tasks;
+
 public abstract class Task {
     public String value;
     public boolean done;
